requirements.txt
dnspython===1.15.0
requests==2.12.1